Endpoint de remoção de turmas
