+++ sme.gclass.application/Interfaces/CasosDeUso/Gsa/Sincronizacao/FormacaoCidade/IIniciarSincronizacaoGsaFormacaoCidadeTurmasExcluirUseCase.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace SME.GoogleClassroom.Aplicacao
+{
+    public interface IIniciarSincronizacaoGsaFormacaoCidadeTurmasExcluirUseCase
+    {
+        Task<bool> Executar(long[] cursosIds);
+    }
+}
